Sessions added

+++ src/main/java/action/DeleteReservationAction.java
@@ -31,11 +31,11 @@ public class DeleteReservationAction {
         this.reservationId = reservationId;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 }
